iio:ad9361: Fix debugfs_entry overflow
The last commit did add new entries to the debugfs_entry array, but did not
increase the size resulting in an array overflow and overwriting the fields in
the struct that follow after the array.

While we are at it also add safety checks to prevent this in the future.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>