Use gpio hog to enable sysref generator
Currently the enable pin of the sysref generator core is handled by the
JESD204 PHY driver. From a conceptual point of view this isn't quite right
since the SYSREF signal is placed at the link level and not at the PHY
level.

All the driver does at the moment is to configure the pin as a high-level
output. The same can be achieved using a GPIO hog, so use that instead
until there is a driver for the SYSREF generator.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>