iio: axi_adxcvr: Recalc lane rate regardless of whether div40 clock is provided
In the current implementation the axi_adxcvr driver only calculates the
lane rate by reading the PLL configuration from the hardware if the div40
clock was provided. Otherwise a cached value of the previously set rate is
used.

The div40 clock is not in any way required for computing the PLL settings
though, so remove the check and unconditionally calculate the rate from the
hardware settings.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>