iio: ad9371: Don't expect deframer errors during link setup
Bit 6 of the deframer status is the so called 'deframer IRQ' bit. The bit
gets set when some kind of error occured during link setup. The current
code expects this bit to be set and prints a warning otherwise. This should
not be the case so change the condition to expect that the bit is cleared.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>