zynq-zc706-adv7511-ad9625-fmcadc3.dts: Reuse FMCADC2 devicetree
FMCADC2 and FMCADC3 are identical expect for additional power amplifiers on
the FMCADC3. They both also use an identical HDL design. This means the
devicetrees are also largely identical with the exception of the additional
entries for the power amplifiers.

Instead of having two devicetrees with the same content include the FMCADC2
dts file in the FMCADC3 dts and add the additional power amplifier entries
on-top of it. This mean future changes to the base design only need to be
done in one file.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>