iio: altera_adxcvr: Prevent concurrent access to the same register space
The way the transceiver register maps are exposed by the system design
makes it look as if each logical transceiver has its own independent
register map.

In reality though the register map between logical TX and RX channels that
share the same physical transceiver are the same register map which is
mapped at different addresses.

Concurrent access to these maps can result in undefined behaviour. The
driver needs to make sure that this can not happen.

Since the driver does not know which two transceivers share a register
space introduce a global mutex that prevents concurrent access between any
two transceivers. This is a bit over restrictive, but good enough for now
since there usually shouldn't be any contention for accessing the
transceiver registers.

Long term a better system design is required that does not hide the fact
that the register maps are shared.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>