iio: altera_adxcvr: Add AVMM bus arbitration helper functions
Acquiring and releasing the AVMM bus arbitration is the same for all
transceiver components. Factor this out into a common helper function
instead of having the same code in multiple places.

Also remove the read-modify-write cycle of the arbitration register. The
mask is 0xff and all bits are going to be overwritten, so no need to
read back the current value.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>