iio: altera_adxcvr: Establish explicit probe ordering to reference clock
The transceiver needs the reference clock to be available and enabled for
initialization and calibration to work correctly. Currently the driver waits a
specified amount of time before it tries to run calibration and hopes that
everything is ready by then.

Rework things to establish a explicit dependency on the reference clock.
This will make sure that the clock is available and correctly configured
before calibration is being run.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>