iio: altera_adxcvr: Add support for configuring the device clock
The device clock needs to run at exactly 1/40 of the lane rate for things
to function correctly. Add support for configuring the clock when the lane
rate changes.

Things are a bit tricky since it is not possible to set the rate of a clock
from within the set_rate() callback of another clock. So a delayed work
item is used to change the device clock rate. This opens up the
possibility that another set_rate() operation starts before the device
clock has been successfully update. This requires special care to avoid
race conditions.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>