iio: altera_adxcvr: Ensure arbitration is acquired
Arbitration is not necessarily granted immediately. If register access is
done to the transceiver register space when the arbitration has not been granted
yet register reads will return bogus data and writes will be lost. This can
leave the transceiver in an invalid configuration and cause undefined
behaviour.

So make sure that the arbitration has actually been granted by checking the
AVMM_BUSY flag in that status register.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>